size

+++ 0x00-hello_world/6-size.c
@@ -2,6 +2,7 @@
 
 /**
 * This program prints size of data types
+* Return is 0
 */
 int main(void)
 {
@@ -10,10 +11,13 @@ int c;
 int d;
 float e;
 
-printf("size of a char: %lu.\n", sizeof(a));
-printf("size of an int: %lu.\n", sizeof(b));
-printf("size of a long int: %lu.\n", sizeof(c))
-printf("size of a long long int: %lu.\n", sizeof(d));
+/** 
+* Prints sizes of data types
+*/
+printf("size of a char: %lu.\n", (short)sizeof(a));
+printf("size of an int: %lu.\n", (short)sizeof(b));
+printf("size of a long int: %lu.\n", (unsigned long)sizeof(c))
+printf("size of a long long int: %lu.\n", (unsigned long long)sizeof(d));
 
-return(0);
+return (0);
 }
